--- src/components/AboutMe/css/index.css
@@ -1,35 +1,0 @@
-.centered {
-    min-height: 80vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-.centeredPhoto{
-    min-height: 80vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-.hh1{
-    font-size: 100px;
-    display:contents;
-    padding: 0;
-    line-height: 40px;
-
-}
-
-p{
-    font-size: 20px;
-    display: flex;
-    margin-left: 7px;
-    font-weight: bold;
-    line-height: 0px;
-}
-
-.emailFont{
-    font-size: 17px;
-    margin-left: 14px;
-}
